JS: Share webpack.config.js between clients
Introduce a module in js/webpack-config and use it in the webpack.config.js
