+++ clients/crunchbase/ui/package.json
@@ -1,8 +1,9 @@
 {
-  "name": "test-nvd3",
+  "name": "noms-crunchbase-ui",
   "dependencies": {
     "d3": "^3.4.4",
     "noms": "file:../../../js",
+    "noms-webpack-config": "file:../../../js/webpack-config",
     "nvd3": "^1.8.1",
     "react": "^0.14.1",
     "react-dom": "^0.14.1"
@@ -11,7 +12,6 @@
     "babel-cli": "^6.3.15",
     "babel-core": "^6.3.15",
     "babel-eslint": "^4.1.5",
-    "babel-loader": "^6.2.1",
     "babel-plugin-syntax-async-functions": "^6.3.13",
     "babel-plugin-syntax-flow": "^6.3.13",
     "babel-plugin-transform-async-to-generator": "^6.3.13",
@@ -25,8 +25,7 @@
     "eslint": "^1.9.0",
     "eslint-plugin-react": "^3.8.0",
     "flow-bin": "^0.19.1",
-    "http-server": "^0.8.5",
-    "webpack": "^1.12.11"
+    "http-server": "^0.8.5"
   },
   "scripts": {
     "start": "./.npm_build_helper.sh development",
