+++ js/package.json
@@ -14,7 +14,6 @@
     "babel-core": "^6.3.15",
     "babel-eslint": "^4.1.5",
     "babel-generator": "^6.4.2",
-    "babel-loader": "^6.2.1",
     "babel-plugin-syntax-async-functions": "^6.3.13",
     "babel-plugin-syntax-flow": "^6.3.13",
     "babel-plugin-transform-async-to-generator": "^6.3.13",
@@ -27,8 +26,7 @@
     "babel-preset-react": "^6.3.13",
     "eslint": "^1.9.0",
     "eslint-plugin-react": "^3.8.0",
-    "flow-bin": "^0.19.1",
-    "webpack": "^1.12.11"
+    "flow-bin": "^0.19.1"
   },
   "scripts": {
     "pretest": "eslint src/ && flow src/",
