+++ js/webpack-config/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "noms-webpack-config",
+  "version": "0.0.1",
+  "description": "Shared webpack config for noms",
+  "main": "index.js",
+  "dependencies": {
+    "babel-loader": "^6.2.1",
+    "webpack": "^1.12.11"
+  }
+}
